Include the standard headers used directly, not transitively
libc++ pulls in <cstdio>, <string> and <filesystem> through other
headers; libstdc++ and MSVC do not. Every previous CI failure on this
project was some version of "works only because of my machine", so
these are audited rather than discovered on the runner.

Also silences the last -Wdouble-promotion warning in the CLI. [skip ci]

+++ tools/cli/main.cpp
@@ -134,8 +134,8 @@ int info(std::string_view path) {
 
     if (const auto& rg = props.replayGain; !rg.empty()) {
         std::fputs("replaygain:  ", stdout);
-        if (rg.trackGain) std::printf(" track %+.2f dB", *rg.trackGain);
-        if (rg.albumGain) std::printf(" album %+.2f dB", *rg.albumGain);
+        if (rg.trackGain) std::printf(" track %+.2f dB", static_cast<double>(*rg.trackGain));
+        if (rg.albumGain) std::printf(" album %+.2f dB", static_cast<double>(*rg.albumGain));
         std::putchar('\n');
     }
